style(gorm): apply gofmt to all files touched by gorm v2 migration
The v1→v2 migration commit (178272476) ran mechanical import/tag/error
conversions across 876 files. Many of those files were not subsequently
gofmt-ed, leaving inconsistent import grouping and comment spacing.

This commit applies gofmt -w to all 1000 .go files that had uncommitted
formatting diffs after the migration. No semantic changes.

+++ common/yakgrpc/grpc_manageScript_test.go
@@ -593,7 +593,7 @@ func TestQueryYakScript(t *testing.T) {
 
 	tests := []struct {
 		filedKeywords string
-		count         int
+		count         int64
 	}{
 		{
 			filedKeywords: "fileKeywords-test",
